requirements.txt
openai
pdfplumber==0.10.1
pyyaml==6.0
